calculate a recommended NDR for users
Used the default model, Bambu now tried to estimate the number of missing reference annotations in order to suggest an appropriate NDR to use if the user is unaware how good their reference annotations are.
If the user inputs an NDR that is more that 0.1 different from the recommended there will be a warning.