Fix bug in runs where annotation gene names had "gene"
Previously bambu would call genes novel if it read "gene" in the gene name indicating it was assigned by Bambu. This caused errors in reference annotations that had gene in the gene name causing all genes to be classed as novel and no training being possible. 
Now the assignGeneIds function returns a list of the novel genes instead.